requirements.txt
Flask==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
Flask-WTF==0.15.1
WTForms==2.3.3
python-dotenv==0.19.0
Flask-Migrate==3.1.0
Flask-Bcrypt==0.7.1
email-validator==1.1.3
cryptography==36.0.0
